aula 3

+++ ListaBasica/exercicio7.py
@@ -1,5 +1,5 @@
 # 7. Calculadora de Quatro Operações:
-num1 = int(input("Dê um numero: \n"))
-num2 = int(input("Dê outro numero: \n"))
+num1 = float(input("Dê um numero: \n"))
+num2 = float(input("Dê outro numero: \n"))
 
 print(f"Soma: {num1 + num2} \nDivisao: {num1 / num2} \nMultiplicação: {num1 * num2} \nSubtração: {num1 - num2}")
